Crate rustproof

Rustproof is a compiler plugin for the Rust programming language. It generates verification conditions for functions with supplied preconditions(P) and postconditions(Q). That is, given a supplied postcondition on a function, rustproof uses predicate transformer semantics to generate a weakest precondition(WP). The verification condition P->WP is then checked for satisfiability by a SMT solver (z3). This process results in a proof of function correctness.

Access the rustproof README here.

The following documentation is intended for the devlopers of rustproof. These descibed modules are not user-facing.

Modules

expression
Internal data structure and related functions to represent logical expressions.

parser
Parses rustproof attribute definitions.

reporting
Prints user-facing errors and warnings.

smt_output
Interface between rustproof and libsmt(z3).

weakest_precondition
Generates a weakest precondition from MIR basic block data.

Structs

MirData
Represents the data from the MIR pass relevant to the function being analyzed

MirVisitor

Functions

registrar