Module rustproof::expression

Internal data structure and related functions to represent logical expressions.

Structs

BinaryExpressionData

SignedBitVectorData

UnaryExpressionData

UnsignedBitVectorData

VariableMappingData

Enums

BinaryOperator

Expression

Types

UnaryOperator

Functions

determine_evaluation_type
Recurses through an Expression and returns the type it would evaluate to.

is_signed_type
Checks if a Types, like the return from determine_evaluation_type(), matches one of the supported signed integer types

is_unsigned_type
Checks if a Types, like the return from determine_evaluation_type(), matches one of the supported unsigned integer types

is_valid_signed
Checks if an Expression matches one of the supported signed integer types

is_valid_unsigned
Checks if an Expression matches one of the supported unsigned integer types

string_to_type
Returns a Types interpreted from a String representing a variable or literal type

substitute_variable_with_expression
Recurses through an Expression and replaces any instance of a variable with a given Expression.

ty_check
Recurses through an Expression and checks for validity of types, operands, and integer bounds.