Function rustproof::parser::parse_attribute

pub fn parse_attribute(pre_string: &mut String, post_string: &mut String, attr: &Spanned<Attribute_>)

Analyzes an attribute on a function in the compiled code, and if the attribute is "condition", ensures correct usage. If usage is correct, it stores the argument strings.

Arguments:

pre_string - Empty string. Will contain a user-submitted precondition if found.

post_string - Empty string. Will contain a user-submitted postcondition if found.

attr - The attribute being analyzed.

Remarks:

Currently supported ConstInt: i8, i16, i32, i64, u8, u16, u32, u64, bool