Module rustproof::weakest_precondition

Reexports

extern crate rustc_const_math;

use super::dev_tools;

use super::DEBUG;

use std::process;

use expression::*;

use rustc::mir::repr::*;

use rustc::middle::const_val::ConstVal;

use rustc_const_math::ConstInt;

use rustc_data_structures::indexed_vec::Idx;

use rustc::ty::{Ty, TypeVariants};

use std::rt::begin_panic_fmt;

use term;

Functions

add_overflow
Generates a version of wp "And"ed together with a conditional expression that mimics a check for overflow for the type of var.

add_underflow
Generates a version of wp "And"ed together with a conditional expression that mimics a check for underflow for the type of var.

add_zero_check
Generates a version of wp "And"ed together with a conditional expression that mimics a check to ensure division by 0 does not occur.

gen
Computes the weakest precondition for a given post-condition and series of statments over one or more BasicBlocks from MIR.

gen_expression
Generates an appropriate Expression based on whatever is found as an operand, either a literal or some kind of variable/temp/field

gen_lvalue
Generates an appropriate variable mapping based on whatever variable, temp, or field is found

gen_stmt
Returns a (possibly) modified weakest precondition based on the content of a statement

gen_ty
Returns the type of an operand as a String