Function rustproof::weakest_precondition::add_zero_check

pub fn add_zero_check(wp: &Expression, exp: &Expression) -> Expression

Generates a version of wp "And"ed together with a conditional expression that mimics a check to ensure division by 0 does not occur.

Arguments:

wp - The current weakest precondition that the "div by 0" is to be "And"ed to

exp - The expression to check to make sure it is not divided by 0

Return Value:

Returns the modified weakest precondition with "div by 0" Expression "And"ed

Remarks:

Current supported ConstInt: I8, I16, I32, I64, U8, U16, U32, U64