Function rustproof::weakest_precondition::gen

pub fn gen(index: usize, mir_data: &(Vec<&ArgDecl>, Vec<&BasicBlockData>, Vec<&TempDecl>, Vec<&VarDecl>, String), post_expr: &Option<Expression>) -> Option<Expression>

Computes the weakest precondition for a given post-condition and series of statments over one or more BasicBlocks from MIR.

Arguments:

index - Index of the ArgDecl,TempDecl, OR VarDecl within the respective Vec<>

mir_data - Contains the BasicBlockData and all of the ArgDecls, TempDecls, and VarDecls from the MIR pass

post_expr - Contains the post-condition in expression form

Return Value:

returns the weakest precondtion in the form of an expression within an Option. This is sent to FIXME

Remarks:

This is the main generator for the weakest precondtion and is called to generate the weakest precondition

Debug:

to find debugging statments, grep for "DEBUG PURPOSE:"