Module rustproof::weakest_precondition

Generates a weakest precondition from MIR basic block data.

Modules

overflow
Functions to generate overflow checks in the weakest precondition.

Functions

add_zero_check
Generates a version of wp "And"ed together with a conditional expression that mimics a check to ensure division by 0 does not occur.

gen
Computes the weakest precondition for a given postcondition and a series of statements over one or more MIR basic blocks.

gen_expression
Generates an Expression based on some operand, either a literal or some kind of variable, temp, or field

gen_lvalue
Generates an appropriate variable mapping based on whatever variable, temp, or field is found

gen_stmt
Returns a (possibly) modified weakest precondition based on the content of a statement

gen_ty
Returns the type of an operand as a Types